+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Sort the settings. Fixes STCOR-286.
 * Load region-specific translations if available. Fixes STCOR-261.
 * Updated webpack config to disable CSS variable preservation. Fixes STCOR-260.
+* Provide default translations from `en.json` to all localizations. Fixes STCOR-310.
 
 ## [2.17.1](https://github.com/folio-org/stripes-core/tree/v2.17.1) (2018-12-21)
 [Full Changelog](https://github.com/folio-org/stripes-core/compare/v2.17.0...v2.17.1)